runnablearchs: Fix for x86_64h on 10.9.
OSX 10.9 (the first OS to support Haswell CPUs) falsely considers
x86_64h code to be runnable on all x86_64 machines, even with
pre-Haswell CPUs.  To fix that, in the x86_64h case, we add a MOVBE
instruction to the test program.  This is one of the instructions
added on Haswell, and causes an illegal instruction crash on
pre-Haswell CPUs.

This bug never actually affected legacy-support, since the latter
makes no use of Haswell features, and a build for x86_64h would run
just fine on non-Haswell CPUs.  But we fix it for maximum correctness,
and in case this code is ever reused for another purpose that cares
about this.

This bug only applies to 10.9, since earlier OSes don't support
Haswell at all, and later OSes restrict the runnability appropriately.

TESTED:
Now reports correct archs in all cases, including under 10.9 on both
pre-Haswell and post-Haswell machines.